LinkMaster-Pro/requirements.txt
flask>=2.0.1
requests>=2.31.0
flask-cors>=4.0.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
pyinstaller>=5.13.0
pillow>=10.0.0 